+++ creme/naive_bayes/__init__.py
@@ -2,12 +2,14 @@
 Naive Bayes algorithms.
 """
 from .bernoulli import BernoulliNB
+from .complement import ComplementNB
 from .gaussian import GaussianNB
 from .multinomial import MultinomialNB
 
 
 __all__ = [
     'BernoulliNB',
+    'ComplementNB',
     'GaussianNB',
     'MultinomialNB'
 ]