+++ creme/naive_bayes/bernoulli.py
@@ -123,9 +123,9 @@ class BernoulliNB(base.BaseNB):
             c: math.log(self.p_class(c)) + sum(map(
                 math.log,
                 (
-                    self.p_feature_given_class(f, c)
+                    10e-10 + self.p_feature_given_class(f, c)
                     if f in x and x[f] > self.true_threshold
-                    else 1. - self.p_feature_given_class(f, c)
+                    else 10e-10 + (1. - self.p_feature_given_class(f, c))
                     for f in self.feature_counts
                 )
             ))